Moved re-usable migration classes to their own namespace
