+++ src/MigrateRelationshipsCliCommand.php
@@ -8,6 +8,7 @@ use Dhii\I18n\StringTranslatorConsumingTrait;
 use Exception;
 use Inpsyde\MultilingualPress2to3\Cli\CliUtilsTrait;
 use Inpsyde\MultilingualPress2to3\Db\DatabaseWpdbTrait;
+use Inpsyde\MultilingualPress2to3\Migration\ContentRelationshipMigrator;
 use Throwable;
 use wpdb as Wpdb;
